+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 * More `RDD` method bindings: `randomSplit`, `zipWithUniqueId`,
   `reduceByKey`, `subtractByKey`.
 
+### Changed
+
+* Use inline-java for PairRDD bindings under the hood.
+
 ## [0.7.2] - 2017-12-25
 
 ### Added